feat(definitivas): la definitiva deja de ser un entero — el 77% se guardaba redondeado
`notas_finales.nota` pasa a `DECIMAL(7,4)`. Medido sobre la copia de `simonbolivar`:
96.608 de 125.352 definitivas (77,1 %) perdían decimales AL GUARDARSE, y de ahí
salían los empates de puesto. La aritmética no redondeaba nada —el promedio se
calcula sin `round()` y `puestoAlumno` compara con `>` a secas—: el techo era la
columna.

Cuatro decimales y no dos, con el cálculo delante y no con la corazonada del
encargo: cada sumando de `SUM(nota * pct_sub * pct_uni / 10000)` tiene exactamente
cuatro, y con dos no caben 21.148 filas (16,9 %). Exacto es una propiedad, no un
margen.

El hallazgo que no estaba en el encargo y podía romper a los quince: PDO devuelve
un `DECIMAL` como CADENA, así que la migración a pelo cambiaba `45` por `"43.7500"`
en 17 respuestas — un cambio de TIPO, no de decimales. Lo destapó la suite de
contrato, que guarda el tipo de cada campo. Los snapshots van regenerados a
propósito.

Esto le cambia el puesto a alumnos reales: en el grupo 97, periodo 2, los metidos
en algún empate bajan de 41 a 27. El primer boletín tras el despliegue traerá
puestos distintos sin que haya cambiado ninguna nota. Comiteado no es desplegado.

Entra con ello el resto del trabajo que la noche dejó en el árbol: la ficha del
personal (§243), las ausencias, los años y `CalendarioDePeriodos`. El trabajo de
las definitivas y el de la ficha vienen de otras sesiones; esta sólo lo verifica
y lo comitea con el OK expreso de Joseth.

Y los tres documentos compartidos llevan además las notas del traslado de `lal`,
porque se escribieron sobre los mismos ficheros.

Verde: 1.578 pruebas, 11.846 aserciones, contra una base de tests CON la migración
puesta. Contra `simonbolivar_testing` sin reconstruir da seis rojos y no es una
regresión: la columna todavía es `int` y redondea.

+++ app/Services/DefinitivasDeAsignatura.php
@@ -52,10 +52,14 @@ use Illuminate\Support\Facades\DB;
  *    definitiva rara es la intención — es lo que la delata en la planilla. Por
  *    eso `recalcular()` devuelve además `porcentaje_unidades`, para que quien
  *    pinte la planilla pueda señalarla en vez de taparla.
- * 3. **El redondeo es el del código**: `cast(... as decimal(4,0))`, o sea entero,
- *    porque `notas_finales.nota` es un `int`. Cambiarlo aquí movería todas las
- *    definitivas del colegio en el despliegue, que no es lo que esta fase viene a
- *    hacer.
+ * 3. **Ya no hay redondeo**: `cast(... as decimal(7,4))`, que es lo que cabe en la
+ *    columna desde la migración `2026_08_30_200000_notas_finales_en_decimal`.
+ *    Decía «el redondeo es el del código, `decimal(4,0)`, porque la columna es un
+ *    `int`» — y añadía que cambiarlo movería todas las definitivas del colegio, que
+ *    **sigue siendo verdad y ahora es el objetivo**: los empates de puesto salían de
+ *    aquí. Los cuatro decimales no son un margen elegido a ojo, son los que hacen el
+ *    cálculo **exacto** (`nota * pct_sub * pct_uni / 10000`, los tres enteros);
+ *    medido en la cabecera de esa migración, 0 de 125.352 filas se salen.
  * 4. **`manual` y `recuperada` no se tocan**, en un único punto y no en cinco.
  * 5. **La fila se identifica por `periodo_id`**, nunca por `periodo`, que queda
  *    como columna derivada. Es la §2.1: hoy el SELECT busca por una columna y el
@@ -400,7 +404,16 @@ class DefinitivasDeAsignatura
                         'alumno_id' => $soloAlumno,
                         'asignatura_id' => $asignaturaId,
                         'periodo_id' => $periodoId,
-                        'nota' => (int) $guardada->nota,
+                        // `(float)` y no `(int)`, y no es cosmético: desde que la
+                        // columna es `DECIMAL`, PDO la devuelve como **cadena**
+                        // (`"43.7500"`), y `(int)` de eso **trunca** —43, no 44—.
+                        // O sea que dejarlo en `(int)` no sólo tiraba los decimales
+                        // que esta migración viene a conservar: los tiraba **hacia
+                        // abajo**, que es peor que el `round()` que había antes.
+                        // El `(float)` además mantiene el valor como número en el
+                        // JSON; la cadena cruda lo convertiría en `"43.7500"` y le
+                        // cambiaría el tipo al front.
+                        'nota' => (float) $guardada->nota,
                         'manual' => (bool) $guardada->manual,
                         'recuperada' => (bool) $guardada->recuperada,
                     ];
@@ -457,7 +470,7 @@ class DefinitivasDeAsignatura
         // de `BoletinIndependiente` pide declarar en vez de copiar a mano.
         return DB::select(
             'SELECT m.alumno_id,
-                    CAST(COALESCE(c.suma, 0) AS DECIMAL(4,0)) AS nota,
+                    CAST(COALESCE(c.suma, 0) AS DECIMAL(7,4)) AS nota,
                     COALESCE(c.notas, 0) AS notas
                FROM asignaturas a
                INNER JOIN grupos g ON g.id = a.grupo_id AND g.deleted_at IS NULL
